rename-seo-component
