+++ packages/client/src/pages/ItemDetailPage/ItemDetailPage.tsx
@@ -2,7 +2,7 @@ import React, { useEffect, useState } from "react";
 import { useParams, Link } from "react-router-dom";
 import { fetchDetails } from "../../utils/fetchDetails"; // Function to fetch item details
 import { Hotel, Country, City } from "../../types/models"; // Importing models for type safety
-import SEO from "../../components/SEO/SEO"; // SEO component for dynamic metadata
+import SEO from "../../components/MetaTags/MetaTags"; // SEO component for dynamic metadata
 import "./ItemDetailPage.css";
 
 const ItemDetailPage: React.FC = () => {
